Auto ping redis every X minutes. Add discord private message method in new discord pkg. Fix redis client singleton.
