+++ config/config-dev.json
@@ -24,5 +24,9 @@
     },
     "logger": {
         "level": "debug"
+    },
+    "redis" : {
+        "pool": 1,
+        "pingTimer": 1
     }
 }
