+++ pkg/twitter/stream.go
@@ -2,14 +2,17 @@ package twitter
 
 import (
 	"errors"
+	"github.com/zoulls/provencal-le-gaulois/pkg/utils"
 	"net/url"
+	"time"
 
 	"github.com/ChimeraCoder/anaconda"
 	"github.com/bwmarrin/discordgo"
 
 	"github.com/zoulls/provencal-le-gaulois/config"
+	"github.com/zoulls/provencal-le-gaulois/pkg/discord"
 	"github.com/zoulls/provencal-le-gaulois/pkg/logger"
-	"github.com/zoulls/provencal-le-gaulois/pkg/reply"
+	"github.com/zoulls/provencal-le-gaulois/pkg/redis"
 	"github.com/zoulls/provencal-le-gaulois/pkg/status"
 )
 
@@ -20,7 +23,7 @@ func getAPI() *anaconda.TwitterApi {
 	return anaconda.NewTwitterApi(conf.Twitter.Config.AccessToken, conf.Twitter.Config.AccessTokenSecret)
 }
 
-func StreamTweets(discord *discordgo.Session, sClient *status.Status) {
+func StreamTweets(ds *discordgo.Session, sClient *status.Status, rClient redis.Client) {
 	conf := config.GetConfig()
 	api := getAPI()
 	v := url.Values{}
@@ -28,21 +31,34 @@ func StreamTweets(discord *discordgo.Session, sClient *status.Status) {
 	s := api.PublicStreamFilter(v)
 	defer api.Close()
 
+	lastPing := time.Now()
+
 	for t := range s.C {
+		// Check for status update
 		if conf.StatusUpdate.Enabled {
 			lastStatus, err := sClient.Last(false)
 			if err != nil {
 				logger.Log.Errorf("Error retrieving the last status, %v", err)
 			}
-			err = discord.UpdateStatus(0, lastStatus)
+			err = ds.UpdateStatus(0, lastStatus)
 			if err != nil {
 				logger.Log.Errorf("Error during status update, %v", err)
 			}
 		}
+
+		if utils.MoreThan(conf.Redis.PingTimer, lastPing) {
+			lastPing = time.Now()
+			ping, err := rClient.Ping()
+			if err != nil || ping == nil {
+				logger.Log.Errorf("Error during redis ping, %v", err)
+				discord.SendPrivateMessage(ds, conf.Discord.AdminID, "Error during redis ping")
+			}
+		}
+
 		switch tweet := t.(type) {
 		case anaconda.Tweet:
 			if originalTweet(tweet) {
-				err := createMessage(discord, &tweet)
+				err := createMessage(ds, &tweet)
 				if err != nil {
 					logger.Log.Errorf("Error during send message of tweet, %v", tweet)
 				}
@@ -53,9 +69,9 @@ func StreamTweets(discord *discordgo.Session, sClient *status.Status) {
 	}
 }
 
-func createMessage(discord *discordgo.Session, tweet *anaconda.Tweet) error {
+func createMessage(ds *discordgo.Session, tweet *anaconda.Tweet) error {
 	conf := config.GetConfig()
-	message := reply.FromTweet(tweet)
+	message := discord.FromTweet(tweet)
 	reply := &discordgo.MessageSend{
 		Embed: message,
 	}
@@ -65,7 +81,7 @@ func createMessage(discord *discordgo.Session, tweet *anaconda.Tweet) error {
 		return err
 	}
 
-	_, err = discord.ChannelMessageSendComplex(discordID, reply)
+	_, err = ds.ChannelMessageSendComplex(discordID, reply)
 	return err
 }
